# core/app-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI Agent OS 应用管理平台 API 文档",
        "title": "AI Agent OS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9090",
    "paths": {
...
        "/workspace/api/v1/operate_log/table": {
            "get": {
                "description": "查询 Table 操作日志",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "操作日志"
                ],
                "summary": "查询 Table 操作日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "X-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "租户用户（app 的所有者）",
                        "name": "tenant_user",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "请求用户（实际执行操作的用户）",
                        "name": "request_user",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "应用名",
                        "name": "app",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "完整代码路径",
                        "name": "full_code_path",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "记录ID",
                        "name": "row_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "操作类型：OnTableAddRow, OnTableUpdateRow, OnTableDeleteRows",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码（从1开始）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段（默认：created_at DESC）",
                        "name": "order_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/dto.GetTableOperateLogsResp"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.GetTableOperateLogsResp": {
            "type": "object",
            "properties": {
                "logs": {
                    "description": "日志列表"
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页数量",
                    "type": "integer"
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-Token",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}